# Conflicts:
#	app/src/main/java/me/ccrama/redditslide/Activities/Gallery.java
#	app/src/main/java/me/ccrama/redditslide/Activities/MainActivity.java
#	app/src/main/java/me/ccrama/redditslide/Activities/MultiredditOverview.java
#	app/src/main/java/me/ccrama/redditslide/Activities/Profile.java
#	app/src/main/java/me/ccrama/redditslide/Activities/Search.java
#	app/src/main/java/me/ccrama/redditslide/Activities/SetupWidget.java
#	app/src/main/java/me/ccrama/redditslide/Activities/SubredditView.java
#	app/src/main/java/me/ccrama/redditslide/Activities/Website.java
#	app/src/main/java/me/ccrama/redditslide/Adapters/ContributionPosts.java
#	app/src/main/java/me/ccrama/redditslide/Adapters/SubredditSearchPosts.java
#	app/src/main/java/me/ccrama/redditslide/Fragments/MultiredditView.java
#	app/src/main/java/me/ccrama/redditslide/Fragments/SettingsGeneralFragment.java
#	app/src/main/java/me/ccrama/redditslide/Fragments/SettingsThemeFragment.java
#	app/src/main/java/me/ccrama/redditslide/Fragments/SubmissionsView.java
#	app/src/main/java/me/ccrama/redditslide/Notifications/CheckForMail.java
#	app/src/main/java/me/ccrama/redditslide/Notifications/CheckForMailSingle.java
#	app/src/main/java/me/ccrama/redditslide/Reddit.java
#	app/src/main/java/me/ccrama/redditslide/SettingValues.java
#	app/src/main/java/me/ccrama/redditslide/Views/PeekMediaView.java

+++ app/src/main/java/me/ccrama/redditslide/ForceTouch/PeekView.java
@@ -7,23 +7,12 @@ import android.app.Activity;
 import android.graphics.Color;
 import android.graphics.Point;
 import android.graphics.Rect;
-import android.support.annotation.FloatRange;
-import android.support.annotation.IdRes;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.view.Display;
-import android.view.LayoutInflater;
-import android.view.MotionEvent;
-import android.view.View;
-import android.view.ViewGroup;
+import android.support.annotation.*;
+import android.view.*;
 import android.view.animation.AccelerateDecelerateInterpolator;
 import android.view.animation.DecelerateInterpolator;
 import android.view.animation.Interpolator;
 import android.widget.FrameLayout;
-
-import java.util.HashMap;
-
 import jp.wasabeef.blurry.Blurry;
 import me.ccrama.redditslide.ForceTouch.builder.PeekViewOptions;
 import me.ccrama.redditslide.ForceTouch.callback.OnButtonUp;
@@ -35,7 +24,8 @@ import me.ccrama.redditslide.ForceTouch.util.NavigationUtils;
 import me.ccrama.redditslide.R;
 import me.ccrama.redditslide.Reddit;
 import me.ccrama.redditslide.Views.PeekMediaView;
-import me.ccrama.redditslide.util.LogUtil;
+
+import java.util.HashMap;
 
 public class PeekView extends FrameLayout {
 
@@ -73,20 +63,21 @@ public class PeekView extends FrameLayout {
     private OnPop mOnPop;
 
     int currentHighlight;
-    static int eight = Reddit.dpToPxVertical(8);
+    static final int eight = Reddit.dpToPxVertical(8);
 
     public void highlightMenu(MotionEvent event) {
-        if(currentHighlight != 0){
+        if (currentHighlight != 0) {
             final View v = content.findViewById(currentHighlight);
             Rect outRect = new Rect();
             v.getGlobalVisibleRect(outRect);
-            if(!outRect.contains((int) event.getX(), (int) event.getY())){
+            if (!outRect.contains((int) event.getX(), (int) event.getY())) {
                 currentHighlight = 0;
                 ValueAnimator animator = ValueAnimator.ofInt(eight, eight * 2);
                 animator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
                     @Override
-                    public void onAnimationUpdate(ValueAnimator valueAnimator){
-                        v.setPadding(0,  (Integer) valueAnimator.getAnimatedValue(), 0, (Integer) valueAnimator.getAnimatedValue());
+                    public void onAnimationUpdate(ValueAnimator valueAnimator) {
+                        v.setPadding(0, (Integer) valueAnimator.getAnimatedValue(), 0,
+                                (Integer) valueAnimator.getAnimatedValue());
                     }
                 });
                 animator.setInterpolator(new AccelerateDecelerateInterpolator());
@@ -100,13 +91,14 @@ public class PeekView extends FrameLayout {
             final View v = content.findViewById(i);
             Rect outRect = new Rect();
             v.getGlobalVisibleRect(outRect);
-            if(outRect.contains((int) event.getX(), (int) event.getY()) && i != currentHighlight){
+            if (outRect.contains((int) event.getX(), (int) event.getY()) && i != currentHighlight) {
                 currentHighlight = i;
                 ValueAnimator animator = ValueAnimator.ofInt(eight * 2, eight);
                 animator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
                     @Override
-                    public void onAnimationUpdate(ValueAnimator valueAnimator){
-                        v.setPadding(0,  (Integer) valueAnimator.getAnimatedValue(), 0, (Integer) valueAnimator.getAnimatedValue());
+                    public void onAnimationUpdate(ValueAnimator valueAnimator) {
+                        v.setPadding(0, (Integer) valueAnimator.getAnimatedValue(), 0,
+                                (Integer) valueAnimator.getAnimatedValue());
                     }
                 });
                 animator.setInterpolator(new AccelerateDecelerateInterpolator());
@@ -114,18 +106,17 @@ public class PeekView extends FrameLayout {
                 animator.start();
 
                 break;
-            } else if(outRect.contains((int) event.getX(), (int) event.getY())){
+            } else if (outRect.contains((int) event.getX(), (int) event.getY())) {
                 break;
             }
         }
     }
 
-    public void pop(){
-        if(mOnPop != null)
-        mOnPop.onPop();
+    public void pop() {
+        if (mOnPop != null) mOnPop.onPop();
     }
 
-    public void setOnPop(OnPop mOnPop){
+    public void setOnPop(OnPop mOnPop) {
         this.mOnPop = mOnPop;
     }
 
@@ -135,21 +126,21 @@ public class PeekView extends FrameLayout {
         init(context, options, content, callbacks);
     }
 
-    HashMap<Integer, OnButtonUp> buttons = new HashMap<>();
+    final HashMap<Integer, OnButtonUp> buttons = new HashMap<>();
 
     public void checkButtons(MotionEvent event) {
         for (Integer i : buttons.keySet()) {
             View v = content.findViewById(i);
             Rect outRect = new Rect();
             v.getGlobalVisibleRect(outRect);
-            if(outRect.contains((int) event.getX(), (int) event.getY())){
+            if (outRect.contains((int) event.getX(), (int) event.getY())) {
                 buttons.get(i).onButtonUp();
             }
         }
     }
 
     public void doScroll(MotionEvent event) {
-        ((PeekMediaView)content.findViewById(R.id.peek)).doScroll(event);
+        ((PeekMediaView) content.findViewById(R.id.peek)).doScroll(event);
     }
 
     private void init(Activity context, PeekViewOptions options, @NonNull View content,
@@ -211,7 +202,7 @@ public class PeekView extends FrameLayout {
                         .onto((ViewGroup) androidContentView.getRootView());
 
                 dim.setAlpha(0f);
-            } catch(Exception ignored){
+            } catch (Exception ignored) {
 
             }
         }
@@ -450,11 +441,10 @@ public class PeekView extends FrameLayout {
 
         Blurry.delete((ViewGroup) androidContentView.getRootView());
 
-        if(remove != null)
-            remove.onRemove();
+        if (remove != null) remove.onRemove();
     }
 
-    public void setOnRemoveListener(OnRemove onRemove){
+    public void setOnRemoveListener(OnRemove onRemove) {
         this.remove = onRemove;
     }
